chore: would throw if not in git dir
This was because the `root` value was undefined in exec.js and using the local dir, and if the directory wasn't initialise as a git dir, then the wizard would bail when trying to work out what version of snyk the user had (though, thankfully, only affects dev).

Not marked as a fix as this only affected snyk team working with local version of snyk cli.