feat: --trust-policies implicitly trust deep rules
If a dependency has a polices file that ignores a vuln, `snyk test` will echo out the suggestion, but not ignore the vuln by default. If the user add `snyk test --trust-policies` (and note that this also applies to wizard), it will inherit the deep policy rule and the vuln will be ignored if the path matches.