fix(wizard): don't apply same patch twice
This is when a package is deduped (or npm@3 installed), and the patch tries to apply multiple times to the same filename. So we use the vuln.id + package filename as the key to ensure we only patch once, but then put the list back together so the policy rules are correct at the end of the process.

Ref: SC-965