fix: add policy file to git
Some users are forgetting to add their policy file to git, and then sending up their projects to CI systems like travis, which fail because the policy isn't in place.

This helps mitigate this problem for them. Note that we do not commit the file, I feel that this would be too far.