fix: run sub-process commands inside a shell
Ensures that we execute commands inside a shell and not directly.
This makes it easier to run commands on windows without nasty hacks
(like conditionally appending '.cmd' to the command string.. which
doesn't always work).

We originally used child_process.exec which actually calls
child_process.spawn under the hood with the shell option set to true.

Later we switched to using child_process.spawn directly without
realising that the shell option for spawn defaults to false and
has to be manually set.

Check out the docs:
https://nodejs.org/dist/latest-v6.x/docs/api/child_process.html#child_process_child_process_spawn_command_args_options