fix: use random string when failing to get MAC address for signature
We use a local MAC address to randomise sha1 to sign the message. In case MAC address is unreadable (CI env, for example), we defaulted to a random number, causing a crash in crypto.

Using `Math.random` is probably as random as `uuid`, but `crypto.createHash('sha1').update(mac)` requires a Buffer or a String, and not a number.