fix: do not print patch output unless error
This is causing issues on some users' CI environments due to the
verbosity of `patch`. When things go wrong, we output `patch`s stdout
from as part of the error, which means it gets printed twice (causing
huge logs).

This change means we only print `patch`s stdout on error (which is the
only time its interesting).