feat: allow for error message to be displayed on the cli
If the api response contains the property `cliMessage` the cli will display it.
This will be used to display an error message org being used
with `--org=<orgname>` is not found.

The error that should be displayed is:
```
Org 'missing-org' was not found or you may not have the correct permissions to access the org.
You can use the orgs: org1, org2.
```