chore: do not rely on snyk api access when testing
We used to run `snyk test` on the project as part of our tests. This required authentication with a Snyk token, which was provided as a sensitive env var on Travis. Due to security considerations (see https://docs.travis-ci.com/user/pull-requests/#pull-requests-and-security-restrictions), sensitive env vars are not available during the build on forked PRs. This results in forked PRs always failing on tests, making external contributions much harder than needed.

Since the repo is covered by Snyk tests via our GitHub integration, I consider it to be ok to remove the CLI test on builds, in favour of making external contributions easier.