feat: include the org specified with --org in analytics
Up until now the analytics payload that we send did not include the organisation specified as the --org parameter.
While normal calls to `snyk test` resulted in the right org, analytics was saving data to the user's default organisation.
This change now correctly collects the specified org so that the data is connected correctly.