feat: Fetch policy wasm bundle from S3 and write it to the disk.
We now download and extract the bundle to the .iac-data directory. The url is pointed to a public policy bundle object in S3 for Beta release. We also add a temporary env var to gate the download of the bundle.
This is introduced to allow the smoke tests to skip the stage of downloading a bundle. The smoke tests currently use a fixed subset f the wasm policies, but now that we download new policies on every run, the smoke tests would have different results and would fail.

CC-627