chore: Replace 'js-yaml' library with 'yaml' library for IaC
The current library we are using for parsing (js-yaml) has a few issues and detects both false positives and false negatives. This commit replaces ths with a different library called 'yaml'. This parses YAML contents correctly and ensures our functionality does not break. In this commit, we are adding it to IaC specific files only.

CC-854