fix: Allow grouping of vulns for multiple oss results json
When running snyk test using the --json flag the resulting output
may be a JSON array. This may happen for example, when showing both
os and app results of container tests using the --app-vulns flag or
when using the --all-projects flag for oss tests.
A recent fix allowd grouping for vulns found in each result in the
array json output, but only when --app-vulns flag was on.
This fix allows grouping of array results regardless of the
--app-vulns flag, to align the grouping  behavior we get when there
are multiple results in the array, to the behavior when there is
only a single result.