test: revert to using PATH instead of symlinks
The docker bundle doesn't support symlinks, never did. So I've reverted
the smoke test setup to use PATH as before.

Symlinks aren't supported because it uses "${0}" assuming it'll point
to the bundle directory, but for symlinks, it'll be the symlink
directory. This can be solved using `realpath` but macOS does not
include that command by default.