fix: exclude base image vulns filtering all vulns
A previous change moved the filtering out of base image vulns (based on
the existance of `dockerfileInstruction`) to be before
`dockerfileInstruction` is assigned to the vulns, so all vulns were
filtered out in case we used the `exclude-base-image-vulns` option.

This fix moves the filtering back to where it was before, and
recalculates the value of `ok`, to account for the reason it was moved
in the first place.