fix: support --packageManager override
Add support for projects with custom file name (pip and poetry), users can specify these for snyk test with --file & --packageManager option. If present use the packageManager option to determine which handler load.

This is a bug that packageManager override is not considered, the rest of the code avoids relying on the file name.