chore(ci): merge release and pr pipelines
In order to test our artifacts before a release, we can merge the two pipelines we currently have: PR (builds and runs tests) and Release (builds and releases).

These are currently under the same "workflow" in CircleCI but are split based on a filter which checks if the branch is "master". By removing this filter, the pipelines can be merged. We'll keep the "master" filter as a prerequisite for release.

We now also build once so we can see what was generated and test them. The release job only uploads files.

Lerna is no longer needed as npm workspaces provides better support for this separated build, test, release workflow.

The "should-release" approval step is temporary so that we can manually validate the new pipeline's artifacts are correct on master.