feat: Do not reverse vulns order when grouping
Until today, when grouping vulns using the --group-issues flag,
the "from" and "name" arrays of the grouped vuln information were
returned in the reversed order than what appeared in the non-grouped
output of vulns with the same ID, because each new vulns "from" and
"name" data was pushed in the begining of the array. In addition the
last vuln (of the same ID) "upgradePath" was returned.
In a recent performance improvement of the grouping function, we had
to reverse the vuln array twice, in order to keep this order.
In this change, we suggest to remove the vuln array 2 reverse() calls,
which means that the "from" and "name" arrays of the grouped vulns of
the same ID will appear in the same order as in the non-grouped results.
In addition, the first vuln (of the same ID) upgrade path will be
returned, instead of the last one.