WIP: process driftctl structured output
Example commit to demonstrate how we might start capturing structured
(JSON) driftctl output, so that we can do something (analytics) with it
before returning it to the user.

We could consider adding a new command `driftctl fmt <console|html>`,
that takes driftctl's json structure on stdin and emits the requested
format, to encapsulate formatting logic in one place, instead of having
to duplicate it here. Even if our output format diverges from driftctl's
for console output, we might benefit from not duplicating the HTML logic
in its entirity, instead parameterising it in the driftctl codebase to
accept white-labelling.