fix: Single files scans of non-tf files behind FF
This commit fixes the issue introduced on a previous PR where a single file scan of a non-Terraform file (.tf, .tfvars) would fail as a non-valid IaC file.
This is because we didn't ignore the error when the file has already successfully been parsed in the file take of the parsing.
We will now ignore the error and continue with the scan.