chore: Rename references to v3 api to rest api
The new Snyk rest api was originally released as the v3 api, however
since it contains its own versioning scheme it made no sense to have
another version in each endpoint. Therefore it was renamed to the more
generic "rest api" and the url moved from "/v3" to "/rest". The old url
will continue to work for a while during the transition phase but will
eventually go away.

This patch updates all references from the old v3 naming to the new rest
branding.