chore: fix help text documentation
The support-team recently got a ticket from Yahoo! Japan suggesting our documentation provides
an incorrect example based on associated help text.
What's unintentionally confusing is this section:
>    Default: 4 (the current working directory and 3 sub-directories).

   Example: Limit search to the specified directory (or the current directory if no <PATH> is

  specified) plus three levels of subdirectories.

   --detection-depth=3

The assumption, based on the helptext, is that the example should be `--detection-depth=4`.
It's either that, or something else happening in our code that needs to be taken in account in the help text.

[Ticket in reference](https://snyk.zendesk.com/agent/tickets/22644)