fix: always show path names of invalid IaC files
When an IaC directory scan found no recognised IaC formats one or more
syntantically invalid files of a type we try to parse (yaml, json, HCL),
we'd return the "No valid files found to scan" error. That error is also
shown as a special case for when no recognised config types are found
(but nothing is invalid), as a guard in case users are scanning the
wrong directory.

As a special case of the above, when exactly one invalid file was
present, we'd show a slightly confusing error: "Failed to parse <type>
file", and then the directory path - not the file path. This differs
from the behaviour when some recognised config types have been found,
and all unparseable files are listed, grouped by type.

This commit causes all unparseable file paths to be listed, even when no
recognised config types were found.