fix: return paths for files that errrored (IaC)
This commit adds a filename attribute to the CustomErrors that need to keep the filename(path) information.
Next, if the scan is complete, and we have 0 parsed files but some failed files (>0),  we then throw these errors in an array.
As a last step we destructure the array of errors and get the filename(path) information in the results.

This is a "quick" fix which might make the code uglier, but as we will get rid of the entire area in the upcoming weeks, we do not think it's worth the effort of refactoring it now to fix it properly.

CFG-2026