fix: bump snyk docker plugin version golang fixes
We had issues with Go modules not being detected. Version 5.0.1 of
`snyk-docker-plugin` should fix those issues.
This commit includes a workaround since we have a blocked PR (#3433) for
making app vulns the default using the lates version of
`snyk-docker-plugin`. Until we are able to make scanning app vulns the
default, we need to make sure that we set `exclude-app-vulns` option
when a user doesn't set the `app-vulns` option.