feat: add flag to exclude app vulnerabilities
Although we already have a `--app-vulns` flag, the plan is to enable
app-vulnerability scanning in containers by default. Before doing that,
we want to give customers the opportunity to explicitly opt-out of the
change in the default by introducing a new flag `--exclude-app-vulns`.
This flag will co-exist with the `--app-vulns` flag until it's enabled
by default, at which point the `--app-vulns` flag can be removed.