fix: use printf rather than echo for escaping in cliv2 Makefile
`cliv2/Makefile` is using echo in a way that relies on MacOS/zsh specifics,
where `echo` will interpret escape sequences by default.  This not the case
everywhere. Using `echo -e` works on some platforms (zsh and bash) but it's
not part of POSIX, so using `printf` is the most idiomatic way to do it.