feat: adding support for .NET runtime dependency resolution
This update adds support for users to scan their newer .NET projects (.NET 5+, .NET Core and .NET Standard) projects with the inclusion of the runtime dependencies that will be included in the final build of the given <RuntimeIdentifier> target architecture defined in the `.csproj` files.

Achieve this by adding `--dotnet-runtime-resolution` after the `snyk test` command.

This feature is still experimental and should be as of now not be considered safe for production use.