Chore: Make build script more portable
The current script `cliv2/ scripts/prepare_licenses.sh` assumes it can run python directly on the host.  If someone is using python actively for development that's likely disabled. It's good practice to disable pip from installing in the global namespace using one of:

```
pip config set global.require-virtualenv True
PIP_REQUIRE_VIRTUALENV=1
```

That means when running a build it will fail like so:

```
-- Preparing 3rd Party Licenses
make[1]: *** [/Users/garethr/Documents/cli/cliv2/_cache/prepare-3rd-party-licenses] Error 3
make: *** [build] Error
```

This PR changes the script to use `virtualenv` if it's installed. It will create a venv if needed, or reuse an existing one. If virtualenv is not installed it will fall back on the current behaviour.

This should make the build more portable.