fix: do not purge the code action cache in language server before the action is resolved by the client [IDE-134] (#5062)
This version provides a fix for code action handling in language server. Resolvable Code Actions are working in a two-step modus in language server: First, the action is created with a callback function stored in language server. The second step is the code action resolution. When a client calls language server with the codeaction/resolve request, the stored callback function is looked up in the cache and executed.

Previously, this cache was purged on each code action request call. This is not working well, if code action information is retained by the IDE. Therefore, the cache entry for the callback function is now only purged when the action is resolved.