feat: provide easy way to update snyk-ls in CLI [IDE-144]
This commit just refactors the existent script that covered only go-application-framework, to now support any dependency in the snyk github organization.

It does that by introducing a new optional `--name` parameter that should contain the github repository name.

If not called with this parameter, it will update go-application-framework and snyk-ls.

It should be considered, whether this should be made part of the pipeline or not, e.g.

- execute script
- commit new go.mod/go.sum
- build CLI
- run acceptance tests