chore: log errors in acceptance tests (#5180)
When a Snyk CLI command is expected to exit 0 in acceptance tests, it
would be useful for debugging to see the stderr from the process.

This adds an option to runSnykCLI (logErrors) to do this, opt-in. That
way we don't pollute the log with expected non-zero exits in tests that
expect them.