fix: make dotnet acceptance test resilient to system lib vulns
A dotnet 8.0.0 vulnerability was discovered on 2024-07-09 which caused
dotnet acceptance tests to fail on platforms with the vulnerable version
installed.

While the toolchain could be updated, the test can be made more
robust to future reoccurances of dotnet system vulns.

For this test we care more about the operation of the test command
rather than its findings, so we can test for consistency between exit
code and test output instead.

In updating this test, a defect was discovered in the dotnet plugin
which corrupts json output. The test works around this, but it will
require a dotnet plugin upgrade as a followup to mitigate user impact.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch fix/dotnet-test-fail-20240710
# Changes to be committed:
#	modified:   test/jest/acceptance/snyk-test/basic-test-all-languages.spec.ts
#
# Untracked files:
#	distroless-main-jq.json
#	distroless-main.json
#	distroless-pr.json
#	junit.xml
#	monitor-json-compare.bash
#	snyk-linux-monitor-main-1.jq.json
#	snyk-linux-monitor-main-1.json
#	snyk-linux-monitor-main-2.jq.json
#	snyk-linux-monitor-main-2.json
#	snyk-linux-monitor-main.json
#	snyk-linux-monitor-pr.json
#	snyk-linux-test-main.json
#	test.env
#	test/acceptance/workspaces/swift-app/.build/
#	test/acceptance/workspaces/swift/.build/
#	test/fixtures/find-files/swift/test.build/.build/
#