fix: container support hashing large .jar files
The bumped version of snyk-docker-plugin fixes the exception thrown by the crypto module
`The RangeError: data is too long` by reading and computing the sha1 of .jar files in chunks.
This allows snyk container to scan embedded jar files of up to buffer.constants.MAX_LENGTH.
On 32-bit architectures, this value currently is 230 - 1 (about 1 GiB).
On 64-bit architectures, this value currently is 232 (about 4 GiB).