Adding support for GH Advanced Security SARIF Upload
Github annoucned changes to their SARIF [upload](https://github.blog/changelog/2024-05-06-code-scanning-will-stop-combining-runs-from-a-single-upload/) the impact is when a SARIF that is being uploaded, each run must have unique "category", as defined by GitHub [here](https://docs.github.com/en/code-security/code-scanning/integrating-with-code-scanning/sarif-support-for-code-scanning#runautomationdetails-object)

GH's new requirement presents a new problem for when a file is removed from source because GH will not have an empty result for the previous file to close any previously opened items as would happen with `--all-projects` previously since GH would merge all results into a single result and closed items no longer present, but now that the closure is  based on tool.driver.name + Category and we have unique files, this will be more problematic.

"Open source's" solution is the most obvious, include the targetFile. Since Snyk-iac, is already populating the field with the static value of "snyk-iac".  It seemed wise to combine riff off that and use "snyk-sca" for OpenSource. Then I separated those values with a pipe and included the `targetFile`

This pattern was then applied to "snyk-iac" and added to "snyk-container". I wasn't able to find unit tests for SARIF output for "snyk-iac" or "snyk-container" or flags to test the output; however, given the simliarities in the implementation to snyk-sca I felt it was worth submitting.