Feat: Determine if the project.json is for nuget or NX
The goal was to improve the experience when using `--all-projects`

There may be a better place to put a check like this where an additional read doesn't happen; however, this was the least impactful location I found to make this type of change. If it was done after the  This should allow more use of --all-projects when repositories also have `project.json` file from using [NX/js](https://nx.dev/)

In addition to the extra file read, this solution also creates an additional array `targetFiles_temp` to minimize overall code changes.

In general, if there is a preferred way to approach this instead, I'd be happy to hear that feedback.