fix: oauth refresh interprocess race conditions
Update GAF and LS dependencies with recent fixes for an oauth refresh race
condition that can occur when multiple CLI or LS processes attempt to refresh
the oauth token at the same time.

This is likely to occur when running multiple CLI processes or having multiple
IDE windows open (some IDEs will spawn an LS per window). The clients with a
stale OAuth token will all attempt to use it to make API requests, and the
authenticator will attempt to automatically rotate the tokens. Only one refresh
token is valid at a given time though (with some potential grace period that
we'll ignore because depending on this would be fragile), so without
coordination, one process can "steal" the OAuth refresh from the others,
causing some processes to fail and requiring an interactive re-auth to recover
in some cases.

Relevant PRs included in these dependency updates:

- https://github.com/snyk/go-application-framework/pull/223
- https://github.com/snyk/go-application-framework/pull/225
- https://github.com/snyk/snyk-ls/pull/623

CLI-469