fix: container optional python dependencies
See https://github.com/snyk/snyk-docker-plugin/pull/614

Fixing a bug in snyk-docker-plugin so that optional dependencies are
properly connected in the dep-graph.

This could mean that 'snyk container monitor' commands that previously
timed out or errored may now start working. The fix reduces
unnecessary (optional) paths in the dep-graph and so reduce the work
Snyk needs to do when scanning for vulnerabilities.