feat: dep-graph json output file
When both `--print-deps` and `--json-file-output` are being used
produce a depGraph object in the resulting JSON file.

Does not attempt to fix any existing `--json` output formats to reduce
the potential of breaking changes.

This allows CLI users to see the dependency graph that was resolved
by the plugin code and scanned.