fix: resolve CVE-2025-21614 vulnerability in CLI
Upgrades goproxy to address vulnerability.

As a result of this change within goproxy
https://github.com/elazarl/goproxy/pull/555/commits/a4e2f5c81061935c279356e959cfab9940c7ae24
we uncovered an issue in our tests when running against a server using
an IP address.

This commit adds a line to the `args.ts` file to set the
`NODE_TLS_REJECT_UNAUTHORIZED` environment variable to '0'
when the `--insecure` flag is used. This is necessary to allow
connections to servers with invalid TLS certificates
when the `--insecure` flag is enabled.