fix: stop verifying token on ignore cmd
Ignore command is the sole command (auth aside) that verifies against
Snyk platform that a token is valid.

This is redundant as an invalid token would be identified as invalid on
any requests performed against the platform with a 401 status code.

This change refactors the code to only perform local validation,
verifying that the token is set appropriately either via configuration
or environment variable.