chore: switch ai-bom scans to new service
AI-BOM scanning was initially implemented as a special type of code
scan. This had various drawbacks and let to the development of a new
service to provide more flexibility and allow extending the
functionality in future. This PR switches the backend from
the code engine api (via deeproxy) to the new service ai-bom. The
analysis that is performed is otherwise functionally equivalent.

The fake-server and acceptance tests have been adapted to the new api.