fix: ignore casing fix in the policy lib
We're fixing an issue where where:

You have ignores with different casing e.g. NPM:HAWK:20160119 vs npm:hawk:20160119
The first ignore is expired
The other ignores that are active do not have the same casing
Then no ignores are applied to the vuln
What was happening in the logic is that we were setting the value of the vulnId to the value of whatever matches first when checking the ignore keys. That vulnId is then used to check the ignores, so if you have other ignores that do not have the same casing exactly they are never checked.

This PR introduces logic to have a variable for all matching vuln IDs with all types of casing to then check the ignores against, so every ignore stored against the vuln ID no matter what the casing is will be checked.

IGNR-1568