feat: enable reachability checks in OS extension [OSF-91]
Bumping the version of the OS extension to bring in the check of
reachability setting for tests with the `--reachability` param.

The latest version of the extension also brings in support for the
`monitor` command. This commit relaxes the param validation for
`monitor` in order to make routing via the extension work.