feat(test): Adds support for maven metaversions
Bumps `snyk-mvn-plugin` to `v4.3.0` and adds simple fixture exercising metaversion resolution

Metaversions are versions like `RELEASE` and `LATEST` which will get resolved to a concrete version.
On more recent versions of the `maven-dependency-plugin` these do not get resolved during `dependency:tree` and we require an additional command (`mvn dependency:resolve`) to determine the resolved versions.

This affects `snyk test` and any commands which depend on `snyk test`, notably `snyk sbom` which uses the `print-graph` option for `snyk test` which forces a recent version of the dependency plugin